@@ -1,7 +1,7 @@
 <script lang="ts">
-  export let href;
+  let { href, children } = $props();
 </script>
 
 <a {href} class="link link-primary" target="_blank" rel="noopener">
-  <slot></slot>
+  {@render children?.()}
 </a>
