@@ -1,6 +1,14 @@
+<script lang="ts">
+  interface Props {
+    children?: import("svelte").Snippet;
+  }
+
+  let { children }: Props = $props();
+</script>
+
 <dialog class="modal">
   <div class="modal-box">
-    <slot />
+    {@render children?.()}
   </div>
 
   <form method="dialog" class="modal-backdrop">
