@@ -1,8 +1,13 @@
 <script lang="ts">
   import "../app.css";
   import "@fontsource-variable/outfit";
+  interface Props {
+    children?: import("svelte").Snippet;
+  }
+
+  let { children }: Props = $props();
 </script>
 
 <div class="mx-4 flex flex-col">
-  <slot />
+  {@render children?.()}
 </div>
