@@ -1,8 +1,9 @@
-use std::{fmt, future::Future, pin::Pin};
+use std::{future::Future, pin::Pin};
 
 use anyhow::Result;
 
 use camino::Utf8PathBuf;
+use erpy_types::MessageRole;
 use serde::{Deserialize, Serialize};
 use tokio_stream::Stream;
 
@@ -24,23 +25,6 @@ pub struct Model {
     pub id: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum MessageRole {
-    User,
-    Assistant,
-    System,
-}
-
-impl fmt::Display for MessageRole {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            MessageRole::User => write!(f, "user"),
-            MessageRole::Assistant => write!(f, "assistant"),
-            MessageRole::System => write!(f, "system"),
-        }
-    }
-}
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct MessageHistoryItem {
@@ -85,6 +69,7 @@ pub struct StreamingCompletionResponse {
 #[serde(rename_all = "camelCase")]
 pub struct StreamingCompletionChoice {
     pub delta: DeltaContent,
+    pub finish_reason: Option<String>,
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -111,7 +96,7 @@ impl CompletionResponse {
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct CompletionChoice {
     pub index: i32,
-    pub finish_reason: String,
+    pub finish_reason: Option<String>,
     pub message: CompletionMessage,
 }
 
