@@ -1,8 +1,11 @@
 #![allow(dead_code)]
 
+use std::io::Cursor;
+
 use anyhow::{anyhow, bail, Result};
-use erpy_types::Character;
-use log::info;
+use erpy_types::CharacterInformation;
+use image::imageops::FilterType;
+use log::{debug, info};
 use serde::Deserialize;
 use serde_json::Value;
 
@@ -71,16 +74,25 @@ pub struct CharacterCardData {
     pub tags: Vec<String>,
 }
 
-pub async fn character_from_chub_ai(url: &str) -> Result<Character> {
+pub async fn character_from_chub_ai(url: &str) -> Result<CharacterInformation> {
+    use base64::prelude::*;
+
     let parsed_url = reqwest::Url::parse(url)?;
     let api_url = format!("https://api.chub.ai/api{}?full=true", parsed_url.path());
     let response: ChubAiCharacter = reqwest::get(&api_url).await?.json().await?;
-    info!("received character: {response:#?} for URL {url}");
+    debug!("received character: {response:#?} for URL {url}");
 
     let mut first_messages = vec![response.node.definition.first_message];
     first_messages.extend(response.node.definition.alternate_greetings);
 
-    Ok(Character {
+    let image_bytes = reqwest::get(&response.node.definition.avatar)
+        .await?
+        .bytes()
+        .await?;
+
+    let base64 = BASE64_STANDARD.encode(image_bytes);
+
+    Ok(CharacterInformation {
         name: response.node.definition.name,
         description: response.node.definition.description,
         personality: response.node.definition.personality,
@@ -92,12 +104,15 @@ pub async fn character_from_chub_ai(url: &str) -> Result<Character> {
             response.node.definition.system_prompt
         },
         avatar: Some(response.node.definition.avatar),
+        image_base64: Some(base64),
     })
 }
 
-pub fn character_from_png_bytes(bytes: &[u8]) -> Result<Character> {
+pub fn character_from_png_bytes(bytes: &[u8]) -> Result<CharacterInformation> {
     use base64::prelude::*;
     use zune_png::PngDecoder;
+
+    info!("parsing character data from PNG file");
 
     let mut decoder = PngDecoder::new(&*bytes);
     decoder.decode()?;
@@ -116,12 +131,21 @@ pub fn character_from_png_bytes(bytes: &[u8]) -> Result<Character> {
     let json = BASE64_STANDARD.decode(base64)?;
     let json: CharacterCard = serde_json::from_slice(&json)?;
 
-    // TODO when avatar is "none"/None/not a URL, store it on the server somewhere (not sure how static files work in tauri)
+    info!("parsed data for character '{}'", json.data.name);
+
+    let image = image::load_from_memory(bytes)?;
+    let image = image.resize(500, 500, FilterType::Lanczos3);
+    let image = image.into_rgba8();
+
+    let mut writer = Cursor::new(Vec::new());
+    image.write_to(&mut writer, image::ImageFormat::WebP)?;
+    let base64 = BASE64_STANDARD.encode(writer.get_ref());
+    let image_base64 = format!("data:image/webp;base64,{}", base64);
 
     let mut first_messages = vec![json.data.first_message];
     first_messages.extend(json.data.alternate_greetings);
 
-    Ok(Character {
+    Ok(CharacterInformation {
         name: json.data.name,
         description: json.data.description,
         personality: json.data.personality,
@@ -133,16 +157,17 @@ pub fn character_from_png_bytes(bytes: &[u8]) -> Result<Character> {
         },
         first_messages,
         avatar: json.data.avatar,
+        image_base64: Some(image_base64),
     })
 }
 
-pub async fn character_fron_png_url(url: &str) -> Result<Character> {
+pub async fn character_fron_png_url(url: &str) -> Result<CharacterInformation> {
     let bytes = reqwest::get(url).await?.bytes().await?;
 
     character_from_png_bytes(&*bytes)
 }
 
-pub async fn character_from_string(string: &str) -> Result<Character> {
+pub async fn character_from_string(string: &str) -> Result<CharacterInformation> {
     if string.starts_with("https://") {
         let url = reqwest::Url::parse(string)?;
         if url.host_str() == Some("chub.ai") {
