@@ -1,16 +1,16 @@
-use std::path::PathBuf;
-
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Deserialize, Serialize, Clone)]
+#[derive(Serialize, Deserialize, Debug, Default)]
 #[serde(rename_all = "camelCase")]
-pub struct Config {
-    pub api_url: String,
+pub struct SyncSettings {
+    pub server_url: Option<String>,
+    pub client_id: Option<String>,
     pub api_key: Option<String>,
-    pub user_name: String,
-    pub selected_model_path: Option<String>,
+}
 
-    pub kobold_cpp_path: Option<PathBuf>,
+#[derive(Serialize, Deserialize, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct LlmSettings {
     pub max_tokens: Option<usize>,
     pub temperature: Option<f64>,
     pub frequency_penalty: Option<f32>,
@@ -20,21 +20,25 @@ pub struct Config {
     pub seed: Option<i64>,
 }
 
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            api_url: "http://localhost:1234/v1".into(),
-            api_key: None,
-            user_name: "User".into(),
-            selected_model_path: None,
-            kobold_cpp_path: None,
-            max_tokens: None,
-            temperature: None,
-            frequency_penalty: None,
-            presence_penalty: None,
-            repeat_penalty: None,
-            top_p: None,
-            seed: None,
-        }
-    }
+#[derive(Serialize, Deserialize, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct NotificationSettings {
+    pub new_message: bool,
+}
+
+#[derive(Serialize, Deserialize, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct ExperimentalSettings {
+    pub local_llm: bool,
+    pub text_to_speech: bool,
+}
+
+#[derive(Serialize, Deserialize, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct Config {
+    pub user_name: String,
+    pub notifications: NotificationSettings,
+    pub sync: SyncSettings,
+    pub llm: LlmSettings,
+    pub experimental: ExperimentalSettings,
 }
