@@ -5,12 +5,12 @@ use character::character_from_string;
 use config::Config;
 use erpy_ai::{open_ai::OpenAiCompletions, CompletionApis, CompletionRequest, MessageHistoryItem};
 use erpy_ai::{CompletionApi, ModelInfo};
-use erpy_types::Character;
+use erpy_types::CharacterInformation;
 use erpy_types::Chat;
+use log::debug;
 use log::{info, LevelFilter};
 use serde::{Deserialize, Serialize};
 use tauri::{AppHandle, Emitter, Listener, Manager};
-use tauri_plugin_sql::{Migration, MigrationKind};
 use tokio::sync::{oneshot, Mutex};
 use tokio_stream::StreamExt;
 
@@ -44,6 +44,14 @@ fn list_models_on_disk() -> TAResult<Vec<ModelInfo>> {
     Ok(models)
 }
 
+pub fn estimate_tokens(history: &[MessageHistoryItem]) -> usize {
+    let total_len: usize = history
+        .iter()
+        .fold(0, |count, item| count + item.content.len());
+
+    total_len / 4
+}
+
 #[tauri::command]
 async fn chat_completion(
     app: AppHandle,
@@ -51,9 +59,10 @@ async fn chat_completion(
     message_history: Vec<MessageHistoryItem>,
 ) -> TAResult<()> {
     info!(
-        "received request to chat with history: {:#?}",
-        message_history
+        "received request to chat with {} tokens",
+        estimate_tokens(&message_history)
     );
+    debug!("chat history: {message_history:#?}");
     let state = app.state::<State>();
     let lock = state.completions.lock().await;
     let Some(api) = lock.as_ref() else {
@@ -62,15 +71,15 @@ async fn chat_completion(
 
     let request = CompletionRequest {
         messages: message_history,
-        temperature: config.temperature,
+        temperature: config.llm.temperature,
         model: Default::default(),
         stream: true,
-        max_tokens: config.max_tokens,
-        frequency_penalty: config.frequency_penalty,
-        presence_penalty: config.presence_penalty,
-        repeat_penalty: config.repeat_penalty,
-        top_p: config.top_p,
-        seed: config.seed,
+        max_tokens: config.llm.max_tokens,
+        frequency_penalty: config.llm.frequency_penalty,
+        presence_penalty: config.llm.presence_penalty,
+        repeat_penalty: config.llm.repeat_penalty,
+        top_p: config.llm.top_p,
+        seed: config.llm.seed,
     };
 
     let mut stream = api.get_completions_stream(&request).await?;
@@ -93,6 +102,7 @@ async fn chat_completion(
             .expect("failed to emit completion");
     }
 
+    info!("completion stream finished");
     app.emit("completion_done", ())
         .expect("failed to emit completion-done");
 
@@ -100,14 +110,17 @@ async fn chat_completion(
 }
 
 #[tauri::command]
-async fn fetch_character(character_url: String) -> TAResult<Character> {
+async fn fetch_character(character_url: String) -> TAResult<CharacterInformation> {
+    info!("creating character from URL {character_url}");
     let character = character_from_string(&character_url).await?;
     Ok(character)
 }
 
 #[tauri::command]
-async fn upload_character_pngs(pngs: Vec<String>) -> TAResult<Vec<Character>> {
+async fn upload_character_pngs(pngs: Vec<String>) -> TAResult<Vec<CharacterInformation>> {
     use base64::prelude::*;
+
+    info!("uploading {} character PNG files", pngs.len());
 
     let mut characters = Vec::new();
     for base64 in pngs {
@@ -115,7 +128,7 @@ async fn upload_character_pngs(pngs: Vec<String>) -> TAResult<Vec<Character>> {
             .decode(base64)
             .map_err(|e| anyhow!("failed to decode base64: {e}"))?;
         let character = character_from_png_bytes(&bytes)?;
-        info!("received character: {:#?}", character);
+        debug!("received character: {:#?}", character);
 
         characters.push(character);
     }
@@ -136,12 +149,15 @@ async fn active_model(app: AppHandle) -> Option<String> {
 }
 
 #[tauri::command]
-async fn summarize(app: AppHandle, chat: Chat, config: Config, prompt: String) -> TAResult<String> {
+async fn summarize(app: AppHandle, chat: Chat, prompt: String) -> TAResult<String> {
     let state = app.state::<State>();
-    let api = &state.completions;
-    // let summary = chat::summarize(&chat, &api, &prompt).await?;
-
-    Ok("TODO".into())
+    let lock = state.completions.lock().await;
+    if let Some(api) = lock.as_ref() {
+        let summary = chat::summarize(&chat, &api, &prompt).await?;
+        Ok(summary)
+    } else {
+        bail!("no model loaded")
+    }
 }
 
 #[derive(Serialize, Debug, Clone, Copy)]
@@ -212,6 +228,15 @@ async fn load_model(app: AppHandle, payload: LoadModel) -> TAResult<()> {
     Ok(())
 }
 
+#[tauri::command]
+async fn unload_model(app: AppHandle) -> TAResult<()> {
+    let mutex = app.state::<State>();
+    let mut lock = mutex.completions.lock().await;
+    lock.take();
+
+    Ok(())
+}
+
 #[derive(Debug, Serialize)]
 #[serde(tag = "type", rename_all = "kebab-case")]
 enum ConnectionTestResult {
@@ -223,7 +248,15 @@ enum ConnectionTestResult {
 async fn test_connection(api_url: String, api_key: Option<String>) -> ConnectionTestResult {
     let api = OpenAiCompletions::new(api_url, api_key);
     match api.list_models().await {
-        Ok(_) => ConnectionTestResult::Success,
+        Ok(models) => {
+            if models.is_empty() {
+                ConnectionTestResult::Failure {
+                    error: "No models found".to_string(),
+                }
+            } else {
+                ConnectionTestResult::Success
+            }
+        }
         Err(e) => ConnectionTestResult::Failure {
             error: format!("Failed to connect: {e}"),
         },
@@ -232,43 +265,8 @@ async fn test_connection(api_url: String, api_key: Option<String>) -> Connection
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
-    let migrations = vec![
-        Migration {
-            version: 1,
-            description: "create_initial_tables",
-            sql: "CREATE TABLE characters (
-                    id INTEGER PRIMARY KEY, 
-                    url VARCHAR, 
-                    payload VARCHAR NOT NULL
-                )",
-            kind: MigrationKind::Up,
-        },
-        Migration {
-            version: 2,
-            description: "add_chat_history",
-            sql: "CREATE TABLE chats (
-                    id INTEGER PRIMARY KEY, 
-                    title VARCHAR,
-                    character_id INTEGER NOT NULL REFERENCES characters (id), 
-                    payload VARCHAR NOT NULL
-                )",
-            kind: MigrationKind::Up,
-        },
-        Migration {
-            version: 3,
-            description: "add_config",
-            sql: "CREATE TABLE config (id INTEGER PRIMARY KEY, payload VARCHAR NOT NULL)",
-            kind: MigrationKind::Up,
-        },
-        Migration {
-            version: 4,
-            description: "add_archived_column_to_chats",
-            sql: "ALTER TABLE chats ADD COLUMN archived INTEGER NOT NULL DEFAULT 0",
-            kind: MigrationKind::Up,
-        },
-    ];
-
     tauri::Builder::default()
+        .plugin(tauri_plugin_http::init())
         .setup(|app| {
             app.manage(State {
                 completions: Mutex::new(None),
@@ -285,11 +283,6 @@ pub fn run() {
                 .level(LevelFilter::Info)
                 .build(),
         )
-        .plugin(
-            tauri_plugin_sql::Builder::default()
-                .add_migrations("sqlite:erpy.sqlite3", migrations)
-                .build(),
-        )
         .invoke_handler(tauri::generate_handler![
             chat_completion,
             list_models,
@@ -298,9 +291,10 @@ pub fn run() {
             summarize,
             upload_character_pngs,
             load_model,
+            unload_model,
             test_connection,
             list_models_on_disk,
-            get_backends
+            get_backends,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
