@@ -1,15 +1,32 @@
-import { convertFileSrc } from "@tauri-apps/api/core";
-import type { Character, ChatHistoryItem } from "./database";
+import { MessageRole, type Character, type Chat, type ChatHistoryItem } from "./storage";
+import { DateTime } from "luxon";
 
+const numberFormat = new Intl.NumberFormat("en-US");
 const pluralRules = new Intl.PluralRules("en-US");
 
-export function formatTimestamp(timestamp: number) {
-  const date = new Date(timestamp);
-  return date.toLocaleDateString();
+export const toDateTime = (timestamp: number | Date) =>
+  timestamp instanceof Date ? DateTime.fromJSDate(timestamp) : DateTime.fromMillis(timestamp);
+
+export function formatTimestamp(timestamp: number | Date, format: "long" | "short" = "short") {
+  const date = toDateTime(timestamp);
+  return date.toLocaleString(format === "long" ? DateTime.DATETIME_MED : DateTime.DATETIME_SHORT);
+}
+
+export function formatTime(timestamp: number | Date) {
+  const date = toDateTime(timestamp);
+  return date.toLocaleString(DateTime.TIME_SIMPLE);
+}
+
+export function sqliteDateTime(date: Date): string {
+  return DateTime.fromJSDate(date).toFormat("yyyy-MM-dd HH:mm:ss");
 }
 
 export function clamp(value: number, min: number, max: number) {
   return Math.min(Math.max(value, min), max);
+}
+
+export function formatNumber(n: number): string {
+  return numberFormat.format(n);
 }
 
 export function substituteParams(
@@ -33,9 +50,9 @@ export function getInitialChatHistory(
   modelId: string,
 ): ChatHistoryItem[] {
   const systemPrompts = [
-    character.data.system_prompt,
-    character.data.personality,
-    character.data.description,
+    character.systemPrompt,
+    character.personality,
+    character.description,
   ].filter((s) => s.trim().length > 0);
   let systemMessage = "";
   for (const msg of systemPrompts) {
@@ -44,11 +61,11 @@ export function getInitialChatHistory(
 
   const messages: ChatHistoryItem[] = [
     {
-      role: "system",
+      role: MessageRole.System,
       content: [
         {
           content: systemMessage,
-          timestamp: Date.now(),
+          timestamp: new Date(),
           modelId,
         },
       ],
@@ -57,10 +74,10 @@ export function getInitialChatHistory(
   ];
 
   messages.push({
-    role: "assistant",
-    content: character.data.first_messages.map((content) => ({
+    role: MessageRole.Assistant,
+    content: character.firstMessages.map((content) => ({
       content,
-      timestamp: Date.now(),
+      timestamp: new Date(),
       modelId,
     })),
     chosenAnswer: 0,
@@ -69,7 +86,7 @@ export function getInitialChatHistory(
   return messages.map((message) => ({
     ...message,
     content: message.content.map((content) => ({
-      content: substituteParams(content.content, userName, character.data.name, content.content),
+      content: substituteParams(content.content, userName, character.name, content.content),
       timestamp: content.timestamp,
       modelId,
     })),
@@ -88,18 +105,6 @@ export function pluralize(count: number, singular: string, plural: string) {
   }
 }
 
-export function getAvatar(url: string | undefined): string {
-  if (!url) {
-    return "";
-  }
-
-  if (url.startsWith("http")) {
-    return url;
-  } else {
-    return convertFileSrc(url.replace("asset://", ""));
-  }
-}
-
 export function truncate(str: string | undefined | null, maxLen: number): string | undefined {
   if (str == null) {
     return undefined;
@@ -110,4 +115,12 @@ export function truncate(str: string | undefined | null, maxLen: number): string
   }
 
   return str.slice(0, maxLen - 1) + "…";
+}
+
+export function getChatTitle(chat: Chat): string {
+  return (
+    truncate(chat.title, 40) ||
+    truncate(chat.history.at(2)?.content.at(0)?.content, 40) ||
+    "Untitled"
+  );
 }
