@@ -1,17 +1,11 @@
-import { addCharacters, persistCharacters } from "$lib/database";
-import type { CharacterPayload } from "$lib/types";
+import { type Character, type NewCharacter, type ErpyStorage } from "$lib/storage";
+import type { CharacterInformation } from "$lib/types";
 import { invoke } from "@tauri-apps/api/core";
-import { BaseDirectory, mkdir, writeFile } from "@tauri-apps/plugin-fs";
 
-async function createDirectory(name: string, baseDir: BaseDirectory) {
-  try {
-    await mkdir(name, { baseDir });
-  } catch (error) {
-    // ignored
-  }
-}
-
-export async function createCharactersFromPngs(files: FileList) {
+export async function createCharactersFromPngs(
+  files: FileList,
+  storage: ErpyStorage,
+): Promise<Character[]> {
   const pngs = await Promise.all(
     Array.from(files).map(async (file) => {
       const reader = new FileReader();
@@ -25,35 +19,35 @@ export async function createCharactersFromPngs(files: FileList) {
     }),
   );
 
-  await createDirectory("avatars", BaseDirectory.AppData);
-  const characterPayloads: CharacterPayload[] = await invoke("upload_character_pngs", { pngs });
-  const characters = await persistCharacters(characterPayloads);
+  const characterPayloads: CharacterInformation[] = await invoke("upload_character_pngs", { pngs });
 
-  let i = 0;
-  for (const character of characters) {
-    const file = files.item(i);
-    if (!file) continue;
-    const data = await file.arrayBuffer();
-    const array = new Uint8Array(data);
-    const path = character.data.avatar!.replace("asset://", "");
+  const newCharacters: NewCharacter[] = [];
+  for (let i = 0; i < characterPayloads.length; i++) {
+    const character = characterPayloads[i];
 
-    await writeFile(path, array, {
-      baseDir: BaseDirectory.AppData,
+    newCharacters.push({
+      payload: { ...character },
+      url: null,
+      imageBase64: character.image_base64,
     });
-
-    i += 1;
   }
+
+  return await storage.persistCharacters(newCharacters);
 }
 
-export async function createCharacterFromUrls(urls: string[]) {
-  const characters = await Promise.all(
-    urls.map((url) =>
-      invoke<CharacterPayload>("fetch_character", { characterUrl: url }).then((result) => ({
-        url,
-        payload: result,
-      })),
-    ),
+export async function createCharacterFromUrls(
+  urls: string[],
+  storage: ErpyStorage,
+): Promise<Character[]> {
+  const data = await Promise.all(
+    urls.map((url) => invoke<CharacterInformation>("fetch_character", { characterUrl: url })),
   );
 
-  await addCharacters(characters);
+  return storage.persistCharacters(
+    data.map((character) => ({
+      imageBase64: character.image_base64,
+      payload: character,
+      url: null,
+    })),
+  );
 }
